feat(AppsDropdown): move AppsDropdown into Layout
