+++ packages/orion/components.css
@@ -1,4 +1,3 @@
-@import './src/AppsDropdown/appsDropdown.css';
 @import './src/Button/button.css';
 @import './src/Card/card.css';
 @import './src/Checkbox/checkbox.css';
