--- packages/orion/src/AppsDropdown/appsDropdown.css
@@ -1,67 +1,0 @@
-/**
- * Apps Dropdown Trigger
- */
-.orion.dropdown.apps-dropdown {
-  @apply min-w-40 min-h-40;
-}
-
-.orion.dropdown.apps-dropdown > .text {
-  @apply hidden;
-}
-
-.orion.dropdown.apps-dropdown > .icon {
-  @apply flex items-center justify-center m-auto w-40 h-40 rounded-full text-gray-700;
-}
-
-.orion.dropdown.apps-dropdown > .icon:hover {
-  @apply bg-gray-900-8;
-}
-
-.orion.dropdown.apps-dropdown.active > .icon {
-  @apply bg-gray-900-12;
-}
-
-/**
- * Apps Dropdown Menu
- */
-.orion.apps-dropdown.active > .menu {
-  @apply bg-gray-100 left-0;
-  min-width: 206px;
-  width: fit-content;
-  transform: translateX(calc(-50% + 20px));
-}
-
-/**
- * Apps Dropdown Header
- */
-.orion.apps-dropdown > .menu > .header {
-  @apply relative flex items-center bg-white py-16 px-32 cursor-auto;
-}
-
-.orion.apps-dropdown > .menu > .header:before {
-  @apply absolute rounded-full w-8 h-8 bg-wave-500;
-  content: ' ';
-  left: 16px;
-}
-
-/**
- * Apps Dropdown Item
- */
-.orion.dropdown.apps-dropdown > .menu > .item {
-  @apply flex items-center m-8 rounded py-8 px-24;
-}
-
-.orion.dropdown.apps-dropdown > .menu > .item:hover {
-  @apply bg-gray-900-8;
-}
-
-/**
- * Apps Dropdown Image
- */
-.orion.apps-dropdown .orion.apps-dropdown-image {
-  @apply flex justify-center items-center bg-white overflow-hidden mr-8 w-32 h-32 rounded-full border-gray-400;
-}
-
-.orion.apps-dropdown .orion.apps-dropdown-image > .orion.image {
-  @apply mx-0;
-}
