+++ packages/orion/src/Layout/layout.css
@@ -2,6 +2,9 @@
   @apply bg-gray-50 flex flex-col min-h-screen items-center relative;
 }
 
+/**
+ * Center
+ */
 .orion.layout .layout-center {
   @apply flex justify-center w-full;
 }
@@ -10,6 +13,9 @@
   @apply layout-content;
 }
 
+/**
+ * Topbar
+ */
 .orion.layout .layout-topbar {
   @apply fixed top-0 left-0 z-40;
   @apply flex-shrink-0 items-center h-64 bg-white;
@@ -28,6 +34,77 @@
   @apply flex-1 ml-40;
 }
 
+/**
+ * Main
+ */
 .orion.layout .layout-main {
   @apply flex-1 py-24 mt-64;
+}
+
+/**
+ * Apps Dropdown Trigger
+ */
+.orion.layout .layout-apps-dropdown {
+  @apply min-w-40 min-h-40 mx-8;
+}
+
+.orion.layout .layout-apps-dropdown > .text {
+  @apply hidden;
+}
+
+.orion.layout .layout-apps-dropdown > .icon {
+  @apply flex items-center justify-center m-auto w-40 h-40 rounded-full text-gray-700;
+}
+
+.orion.layout .layout-apps-dropdown > .icon:hover {
+  @apply bg-gray-900-8;
+}
+
+.orion.layout .layout-apps-dropdown.active > .icon {
+  @apply bg-gray-900-12;
+}
+
+/**
+ * Apps Dropdown Menu
+ */
+.orion.layout .layout-apps-dropdown > .menu {
+  @apply bg-gray-100 left-0;
+  min-width: 206px;
+  width: fit-content;
+  transform: translateX(calc(-50% + 20px));
+}
+
+/**
+ * Apps Dropdown Header
+ */
+.orion.layout .layout-apps-dropdown > .menu > .header {
+  @apply relative flex items-center bg-white py-16 px-32 cursor-auto;
+}
+
+.orion.layout .layout-apps-dropdown > .menu > .header:before {
+  @apply absolute rounded-full w-8 h-8 bg-wave-500;
+  content: ' ';
+  left: 16px;
+}
+
+/**
+ * Apps Dropdown Item
+ */
+.orion.layout .layout-apps-dropdown > .menu > .item {
+  @apply flex items-center m-8 rounded py-8 px-24;
+}
+
+.orion.layout .layout-apps-dropdown > .menu > .item:hover {
+  @apply bg-gray-900-8;
+}
+
+/**
+ * Apps Dropdown Image
+ */
+.orion.layout .layout-apps-dropdown-image {
+  @apply flex justify-center items-center bg-white overflow-hidden mr-8 w-32 h-32 rounded-full border-gray-400;
+}
+
+.orion.layout .layout-apps-dropdown-image > .orion.image {
+  @apply mx-0;
 }
